fixing id_URL validate

+++ src/app/Services/UserService.php
@@ -37,20 +37,34 @@ class UserService
         return $this->userRepositories->createUser($data);
     }
 
-    public function getUser(string $id)
+    public function getUser(string $id, $user_id)
     {
+        if($id !== $user_id) {
+            return throw new HttpException(403, "You don't have permission to get this user!");
+        }
+
         return $this->userRepositories->getUser($id);
     }
 
-    public function updateUser(array $data, string $id)
+    public function updateUser(array $data, string $id, $user_id)
     {
-
+        $user = $this->userRepositories->getUser($user_id);
+        
+        if($user->user_id !== $user_id) {
+            return throw new HttpException(403, "You don't have permission to update this user!");
+        }
 
         return $this->userRepositories->updateUser($data, $id);
     }
 
-    public function deleteUser(string $id)
+    public function deleteUser(string $id, $user_id)
     {
+        $user = $this->userRepositories->getUser($user_id);
+
+        if($user->user_id !== $user_id) {
+            return throw new HttpException(403, "You don't have permission to update this user!");
+        }
+
         return $this->userRepositories->deleteUser($id);
     }
 }
